update local qasm simulators to return memory, clean up, remove from_old_style_dict()
