+++ qiskit/backends/aer/qasm_simulator.py
@@ -20,11 +20,13 @@ import platform
 
 import numpy as np
 
-from qiskit.result._utils import copy_qasm_from_qobj_into_result, result_from_old_style_dict
+from qiskit.qobj import Result as QobjResult
+from qiskit.qobj import ExperimentResult as QobjExperimentResult
+from qiskit.result import Result
+from qiskit.result._utils import copy_qasm_from_qobj_into_result
 from qiskit.backends import BaseBackend
 from qiskit.backends.aer.aerjob import AerJob
 from qiskit.qobj import Qobj
-from qiskit.qobj import qobj_to_dict
 
 logger = logging.getLogger(__name__)
 
@@ -46,13 +48,16 @@ class QasmSimulator(BaseBackend):
     """C++ quantum circuit simulator with realistic noise"""
 
     DEFAULT_CONFIGURATION = {
-        'name': 'qasm_simulator',
+        'backend_name': 'qasm_simulator',
+        'backend_version': '1.0',
+        'n_qubits': -1,        
         'url': 'https://github.com/QISKit/qiskit-terra/src/qasm-simulator-cpp',
         'simulator': True,
         'local': True,
-        'description': 'A C++ realistic noise simulator for qobj files',
+        'conditional': True,
+        'description': 'A C++ realistic noise simulator for qasm experiments',
         'coupling_map': 'all-to-all',
-        "basis_gates": 'u0,u1,u2,u3,cx,cz,id,x,y,z,h,s,sdg,t,tdg,rzz,' +
+        'basis_gates': 'u0,u1,u2,u3,cx,cz,id,x,y,z,h,s,sdg,t,tdg,rzz,' +
                        'snapshot,wait,noise,save,load'
     }
 
@@ -76,20 +81,26 @@ class QasmSimulator(BaseBackend):
                                     self._configuration.get('exe', 'default locations'))
 
     def run(self, qobj):
-        """Run a qobj on the backend."""
+        """Run qobj asynchronously.
+        Args:
+            qobj (Qobj): payload of the experiments
+        Returns:
+            AerJob: derived from BaseJob        
+        """
         job_id = str(uuid.uuid4())
         aer_job = AerJob(self, job_id, self._run_job, qobj)
         aer_job.submit()
         return aer_job
 
     def _run_job(self, job_id, qobj):
+        """Run experiments in qobj"""
         self._validate(qobj)
-        result = run(qobj, self._configuration['exe'])
+        result = launch(qobj, self._configuration['exe'])
         result['job_id'] = job_id
         copy_qasm_from_qobj_into_result(qobj, result)
 
-        return result_from_old_style_dict(
-            result, [circuit.header.name for circuit in qobj.experiments])
+        experiment_names = [experiment.header.name for experiment in qobj.experiments]
+        return Result(QobjResult(**result), experiment_names)
 
     def _validate(self, qobj):
         for experiment in qobj.experiments:
@@ -104,10 +115,13 @@ class CliffordSimulator(BaseBackend):
     """"C++ Clifford circuit simulator with realistic noise."""
 
     DEFAULT_CONFIGURATION = {
-        'name': 'clifford_simulator',
-        'url': 'https://github.com/QISKit/qiskit-terra/src/qasm-simulator',
+        'backend_name': 'clifford_simulator',
+        'backend_version': '1.0',
+        'n_qubits': -1,
+        'url': 'https://github.com/QISKit/qiskit-terra/src/qasm-simulator-cpp',
         'simulator': True,
         'local': True,
+        'conditional': True,
         'description': 'A C++ Clifford simulator with approximate noise',
         'coupling_map': 'all-to-all',
         'basis_gates': 'cx,id,x,y,z,h,s,sdg,snapshot,wait,noise,save,load'
@@ -159,11 +173,12 @@ class CliffordSimulator(BaseBackend):
             qobj_dict['config'] = {'simulator': 'clifford'}
 
         qobj = Qobj.from_dict(qobj_dict)
-        result = run(qobj, self._configuration['exe'])
+        result = launch(qobj, self._configuration['exe'])
         result['job_id'] = job_id
+        copy_qasm_from_qobj_into_result(qobj, result)
 
-        return result_from_old_style_dict(
-            result, [circuit.header.name for circuit in qobj.experiments])
+        experiment_names = [experiment.header.name for experiment in qobj.experiments]
+        return Result(QobjResult(**result), experiment_names)
 
     def _validate(self):
         return
@@ -218,13 +233,14 @@ class QASMSimulatorDecoder(json.JSONDecoder):
         return obj
 
 
-def run(qobj, executable):
+def launch(qobj, executable):
     """
-    Run simulation on C++ simulator inside a subprocess.
+    Launch a subprocess and run the C++ simulation inside it.
 
     Args:
         qobj (Qobj): qobj dictionary defining the simulation to run
         executable (string): filename (with path) of the simulator executable
+
     Returns:
         dict: A dict of simulation results
     """
@@ -233,7 +249,7 @@ def run(qobj, executable):
     try:
         with subprocess.Popen([executable, '-'],
                               stdin=PIPE, stdout=PIPE, stderr=PIPE) as proc:
-            cin = json.dumps(qobj_to_dict(qobj, version='0.0.1'),
+            cin = json.dumps(qobj.as_dict(),
                              cls=QASMSimulatorEncoder).encode()
             cout, cerr = proc.communicate(cin)
         if cerr:
@@ -246,125 +262,3 @@ def run(qobj, executable):
         msg = "ERROR: Simulator exe not found at: %s" % executable
         logger.error(msg)
         return {"status": msg, "success": False}
-
-
-def cx_error_matrix(cal_error, zz_error):
-    """
-    Return the coherent error matrix for CR error model of a CNOT gate.
-
-    Args:
-        cal_error (double): calibration error of rotation
-        zz_error (double): ZZ interaction term error
-
-    Returns:
-        numpy.ndarray: A coherent error matrix U_error for the CNOT gate.
-
-    Details:
-
-    The ideal cross-resonsance (CR) gate corresponds to a 2-qubit rotation
-        U_CR_ideal = exp(-1j * (pi/2) * XZ/2)
-
-    where qubit-0 is the control, and qubit-1 is the target. This can be
-    converted to a CNOT gate by single-qubit rotations::
-
-        U_CX = U_L * U_CR_ideal * U_R
-
-    The noisy rotation is implemented as
-        U_CR_noise = exp(-1j * (pi/2 + cal_error) * (XZ + zz_error ZZ)/2)
-
-    The retured error matrix is given by
-        U_error = U_L * U_CR_noise * U_R * U_CX^dagger
-    """
-    # pylint: disable=invalid-name
-    if cal_error == 0 and zz_error == 0:
-        return np.eye(4)
-
-    cx_ideal = np.array([[1, 0, 0, 0],
-                         [0, 0, 0, 1],
-                         [0, 0, 1, 0],
-                         [0, 1, 0, 0]])
-    b = np.sqrt(1.0 + zz_error * zz_error)
-    a = b * (np.pi / 2.0 + cal_error) / 2.0
-    sp = (1.0 + 1j * zz_error) * np.sin(a) / b
-    sm = (1.0 - 1j * zz_error) * np.sin(a) / b
-    c = np.cos(a)
-    cx_noise = np.array([[c + sm, 0, -1j * (c - sm), 0],
-                         [0, 1j * (c - sm), 0, c + sm],
-                         [-1j * (c - sp), 0, c + sp, 0],
-                         [0, c + sp, 0, 1j * (c - sp)]]) / np.sqrt(2)
-    return cx_noise.dot(cx_ideal.conj().T)
-
-
-def x90_error_matrix(cal_error, detuning_error):
-    """
-    Return the coherent error matrix for a X90 rotation gate.
-
-    Args:
-        cal_error (double): calibration error of rotation
-        detuning_error (double): detuning amount for rotation axis error
-
-    Returns:
-        numpy.ndarray: A coherent error matrix U_error for the X90 gate.
-
-    Details:
-
-    The ideal X90 rotation is a pi/2 rotation about the X-axis:
-        U_X90_ideal = exp(-1j (pi/2) X/2)
-
-    The noisy rotation is implemented as
-        U_X90_noise = exp(-1j (pi/2 + cal_error) (cos(d) X + sin(d) Y)/2)
-
-    where d is the detuning_error.
-
-    The retured error matrix is given by
-        U_error = U_X90_noise * U_X90_ideal^dagger
-    """
-    # pylint: disable=invalid-name
-    if cal_error == 0 and detuning_error == 0:
-        return np.eye(2)
-    else:
-        x90_ideal = np.array([[1., -1.j], [-1.j, 1]]) / np.sqrt(2)
-        c = np.cos(0.5 * cal_error)
-        s = np.sin(0.5 * cal_error)
-        gamma = np.exp(-1j * detuning_error)
-        x90_noise = np.array([[c - s, -1j * (c + s) * gamma],
-                              [-1j * (c + s) * np.conj(gamma), c - s]]) / np.sqrt(2)
-    return x90_noise.dot(x90_ideal.conj().T)
-
-
-def _generate_coherent_error_matrix(config):
-    """
-    Generate U_error matrix for CX and X90 gates.
-
-    Args:
-        config (dict): the config of a qobj circuit
-
-    This parses the config for the following noise parameter keys and returns a
-    coherent error matrix for simulation coherent noise::
-
-        * 'CX' gate: 'calibration_error', 'zz_error'
-        * 'X90' gate: 'calibration_error', 'detuning_error'
-    """
-    # pylint: disable=invalid-name
-    if 'noise_params' in config:
-        # Check for CR coherent error parameters
-        if 'CX' in config['noise_params']:
-            noise_cx = config['noise_params']['CX']
-            cal_error = noise_cx.pop('calibration_error', 0)
-            zz_error = noise_cx.pop('zz_error', 0)
-            # Add to current coherent error matrix
-            if not cal_error == 0 or not zz_error == 0:
-                u_error = noise_cx.get('U_error', np.eye(4))
-                u_error = u_error.dot(cx_error_matrix(cal_error, zz_error))
-                config['noise_params']['CX']['U_error'] = u_error
-        # Check for X90 coherent error parameters
-        if 'X90' in config['noise_params']:
-            noise_x90 = config['noise_params']['X90']
-            cal_error = noise_x90.pop('calibration_error', 0)
-            detuning_error = noise_x90.pop('detuning_error', 0)
-            # Add to current coherent error matrix
-            if not cal_error == 0 or not detuning_error == 0:
-                u_error = noise_x90.get('U_error', np.eye(2))
-                u_error = u_error.dot(x90_error_matrix(cal_error,
-                                                       detuning_error))
-                config['noise_params']['X90']['U_error'] = u_error
